fix: user profile activity
