+++ web/components/headers/profile-settings.tsx
@@ -17,7 +17,7 @@ export const ProfileSettingsHeader: FC<IProfileSettingHeader> = (props) => {
           <Breadcrumbs>
             <Breadcrumbs.BreadcrumbItem
               type="text"
-              label="Settings"
+              label="My Profile"
               icon={<Settings className="h-4 w-4 text-custom-text-300" />}
               link={`/me/profile`}
             />
