+++ web/layouts/settings-layout/profile/index.ts
@@ -1,2 +1,3 @@
 export * from "./layout";
+export * from "./settings-sidebar";
 export * from "./sidebar";
